+++ src/apps/Handover/Garden/utility/index.ts
@@ -3,3 +3,4 @@ export * from './getKeyFunctions';
 export * from './handoverItemMapping';
 export * from './mcProgress';
 export * from './sortFunctions';
+export * from './getSubtitleHeader';
